disable wand builder by default
