+++ preview_generator/preview/builder/image__wand.py
@@ -3,7 +3,6 @@
 from io import BytesIO
 import typing
 
-from pdf2image import convert_from_bytes
 from wand.image import Color
 from wand.image import Image as WImage
 import wand.version
@@ -15,22 +14,11 @@ from preview_generator.utils import compute_resize_dims
 from preview_generator.utils import imagemagick_supported_mimes
 
 
-def convert_pdf_to_jpeg(pdf: typing.IO[bytes], preview_size: ImgDims) -> BytesIO:
-
-    pdf_content = pdf.read()
-    images = convert_from_bytes(pdf_content)
-
-    output = BytesIO()
-    for image in images:
-        resize_dims = compute_resize_dims(ImgDims(image.width, image.height), preview_size)
-        resized = image.resize((resize_dims.width, resize_dims.height), resample=True)
-        resized.save(output, format="JPEG")
-
-    output.seek(0, 0)
-    return output
-
-
 class ImagePreviewBuilderWand(ImagePreviewBuilder):
+    """
+    WARNING : This builder is deprecated, prefer ImagePreviewBuilderIMConvert instead.
+    """
+
     MIMETYPES = []  # type: typing.List[str]
 
     @classmethod
